refactor(chat): extract useStickToBottom hook from Conversations
Move inline scroll-to-bottom logic (two refs + useLayoutEffect) into
app/src/hooks/useStickToBottom.ts. Behavior unchanged — same instant/smooth
split keyed on thread change + first render. 536 tests pass.
